refactor: Removed NativeAssignmentRepository
