+++ src/Ingester/AssignmentIngester.php
@@ -26,7 +26,7 @@ use Doctrine\ORM\ORMException;
 use Doctrine\Persistence\Mapping\MappingException;
 use OAT\SimpleRoster\DataTransferObject\AssignmentDtoCollection;
 use OAT\SimpleRoster\Exception\UserNotFoundException;
-use OAT\SimpleRoster\Repository\NativeAssignmentRepository;
+use OAT\SimpleRoster\Repository\AssignmentRepository;
 use OAT\SimpleRoster\Repository\NativeUserRepository;
 use Throwable;
 
@@ -35,10 +35,10 @@ class AssignmentIngester
     /** @var NativeUserRepository */
     private $userRepository;
 
-    /** @var NativeAssignmentRepository */
+    /** @var AssignmentRepository */
     private $assignmentRepository;
 
-    public function __construct(NativeUserRepository $userRepository, NativeAssignmentRepository $assignmentRepository)
+    public function __construct(NativeUserRepository $userRepository, AssignmentRepository $assignmentRepository)
     {
         $this->userRepository = $userRepository;
         $this->assignmentRepository = $assignmentRepository;
@@ -67,6 +67,6 @@ class AssignmentIngester
             $assignment->setUserId($existingUsernames[$assignment->getUsername()]);
         }
 
-        $this->assignmentRepository->insertMultiple($assignments);
+        $this->assignmentRepository->insertMultipleNatively($assignments);
     }
 }
