+++ src/Repository/AssignmentRepository.php
@@ -25,8 +25,12 @@ namespace OAT\SimpleRoster\Repository;
 use DateTime;
 use Doctrine\ORM\EntityNotFoundException;
 use Doctrine\ORM\NonUniqueResultException;
+use Doctrine\ORM\ORMException;
+use Doctrine\ORM\Query\ResultSetMapping;
 use Doctrine\ORM\Tools\Pagination\Paginator;
 use Doctrine\Persistence\ManagerRegistry;
+use Doctrine\Persistence\Mapping\MappingException;
+use OAT\SimpleRoster\DataTransferObject\AssignmentDtoCollection;
 use OAT\SimpleRoster\Entity\Assignment;
 use Symfony\Component\Uid\UuidV6;
 
@@ -86,5 +90,32 @@ class AssignmentRepository extends AbstractRepository
         }
 
         return new Paginator($queryBuilder->getQuery(), false);
+    }
+
+    /**
+     * @throws ORMException
+     * @throws MappingException
+     */
+    public function insertMultipleNatively(AssignmentDtoCollection $assignments): void
+    {
+        $queryParts = [];
+        foreach ($assignments as $assignmentDto) {
+            $queryParts[] = sprintf(
+                "('%s', %s, '%s', '%s', %s)",
+                (string)$assignmentDto->getId(),
+                $assignmentDto->getUserId(),
+                (string)$assignmentDto->getLineItemId(),
+                $assignmentDto->getState(),
+                0
+            );
+        }
+
+        $query = sprintf(
+            'INSERT INTO assignments (id, user_id, line_item_id, state, attempts_count) VALUES %s',
+            implode(',', $queryParts)
+        );
+
+        $this->_em->createNativeQuery($query, new ResultSetMapping())->execute();
+        $this->_em->clear();
     }
 }
