+++ tests/Integration/Repository/AssignmentRepositoryTest.php
@@ -25,6 +25,8 @@ namespace OAT\SimpleRoster\Tests\Integration\Repository;
 use DateInterval;
 use DateTime;
 use Doctrine\ORM\EntityNotFoundException;
+use OAT\SimpleRoster\DataTransferObject\AssignmentDto;
+use OAT\SimpleRoster\DataTransferObject\AssignmentDtoCollection;
 use OAT\SimpleRoster\Entity\Assignment;
 use OAT\SimpleRoster\Repository\AssignmentRepository;
 use OAT\SimpleRoster\Tests\Traits\DatabaseTestingTrait;
@@ -45,13 +47,14 @@ class AssignmentRepositoryTest extends KernelTestCase
         self::bootKernel();
 
         $this->setUpDatabase();
-        $this->loadFixtureByFilename('usersWithStartedButStuckAssignments.yml');
 
         $this->subject = self::$container->get(AssignmentRepository::class);
     }
 
     public function testItCanFindAssignmentById(): void
     {
+        $this->loadFixtureByFilename('usersWithStartedButStuckAssignments.yml');
+
         $assignment = $this->subject->findById(new UuidV6('00000001-0000-6000-0000-000000000000'));
 
         self::assertSame(Assignment::STATE_STARTED, $assignment->getState());
@@ -68,6 +71,8 @@ class AssignmentRepositoryTest extends KernelTestCase
 
     public function testItCanReturnAssignmentsByStateAndUpdatedAt(): void
     {
+        $this->loadFixtureByFilename('usersWithStartedButStuckAssignments.yml');
+
         $dateTime = (new DateTime())->add(new DateInterval('P1D'));
         $assignments = $this->subject->findByStateAndUpdatedAtPaged(Assignment::STATE_STARTED, $dateTime);
 
@@ -77,10 +82,44 @@ class AssignmentRepositoryTest extends KernelTestCase
 
     public function testItCanReturnAssignmentsByStateAndUpdatedAtPaginated(): void
     {
+        $this->loadFixtureByFilename('usersWithStartedButStuckAssignments.yml');
+
         $dateTime = (new DateTime())->add(new DateInterval('P1D'));
         $assignments = $this->subject->findByStateAndUpdatedAtPaged(Assignment::STATE_STARTED, $dateTime, 2, 3);
 
         self::assertCount(3, $assignments->getIterator());
         self::assertCount(9, $assignments);
+    }
+
+    public function testItCanInsertMultipleAssignments(): void
+    {
+        $this->loadFixtureByFilename('usersWithStartedButStuckAssignments.yml');
+
+        $lineItemId = new UuidV6('00000001-0000-6000-0000-000000000000');
+
+        $assignmentId1 = new UuidV6('00000001-0000-6000-0000-000000000000');
+        $assignment1 = new AssignmentDto($assignmentId1, Assignment::STATE_READY, $lineItemId, 'user1', 1);
+
+        $assignmentId2 = new UuidV6('00000002-0000-6000-0000-000000000000');
+        $assignment2 = new AssignmentDto($assignmentId2, Assignment::STATE_READY, $lineItemId, 'user2', 1);
+
+        $assignmentId3 = new UuidV6('00000003-0000-6000-0000-000000000000');
+        $assignment3 = new AssignmentDto($assignmentId3, Assignment::STATE_READY, $lineItemId, 'user3', 1);
+
+        $assignmentCollection = (new AssignmentDtoCollection())
+            ->add($assignment1)
+            ->add($assignment2)
+            ->add($assignment3);
+
+        $this->subject->insertMultipleNatively($assignmentCollection);
+
+        $assignments = $this->subject->findBy(['id' => [(string)$assignmentId1, $assignmentId2, $assignmentId3]]);
+        self::assertCount(3, $assignments);
+
+        /** @var Assignment $assignment */
+        foreach ($assignments as $assignment) {
+            self::assertSame(Assignment::STATE_READY, $assignment->getState());
+            self::assertSame(0, $assignment->getAttemptsCount());
+        }
     }
 }
