+++ tests/Unit/Ingester/AssignmentIngesterTest.php
@@ -27,7 +27,7 @@ use OAT\SimpleRoster\DataTransferObject\AssignmentDtoCollection;
 use OAT\SimpleRoster\Entity\Assignment;
 use OAT\SimpleRoster\Exception\UserNotFoundException;
 use OAT\SimpleRoster\Ingester\AssignmentIngester;
-use OAT\SimpleRoster\Repository\NativeAssignmentRepository;
+use OAT\SimpleRoster\Repository\AssignmentRepository;
 use OAT\SimpleRoster\Repository\NativeUserRepository;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
@@ -38,7 +38,7 @@ class AssignmentIngesterTest extends TestCase
     /** @var NativeUserRepository|MockObject */
     private $userRepository;
 
-    /** @var NativeAssignmentRepository|MockObject */
+    /** @var AssignmentRepository|MockObject */
     private $assignmentRepository;
 
     /** @var AssignmentIngester */
@@ -49,7 +49,7 @@ class AssignmentIngesterTest extends TestCase
         parent::setUp();
 
         $this->userRepository = $this->createMock(NativeUserRepository::class);
-        $this->assignmentRepository = $this->createMock(NativeAssignmentRepository::class);
+        $this->assignmentRepository = $this->createMock(AssignmentRepository::class);
 
         $this->subject = new AssignmentIngester($this->userRepository, $this->assignmentRepository);
     }
@@ -102,7 +102,7 @@ class AssignmentIngesterTest extends TestCase
 
         $this->assignmentRepository
             ->expects(self::once())
-            ->method('insertMultiple')
+            ->method('insertMultipleNatively')
             ->with($assignmentCollection);
 
         $this->subject->ingest($assignmentCollection);
